Job worker
A job worker which loops through institution list and tries to get an
exclusive lock for a institution.  Having a lock it runs a job from the
institution's job queue and goes on to the next institution.  If a job
fails, but is retryable, it is rescheduled.

Multiple instances of this worker can be started.  Locks and queues are
stored in redis.

Note: running LUA scripts on the server guarantees atomic execution.
Source: https://redis.io/docs/manual/programmability/eval-intro/

> Redis guarantees the script's atomic execution. While executing the
> script, all server activities are blocked during its entire
> runtime. These semantics mean that all of the script's effects either
> have yet to happen or had already happened.