Job worker
A job worker which loops through institution list and tries to get an
exclusive lock for a institution.  Having a lock it runs a job from the
institution's job queue and goes on to the next institution.  If a job
fails, but is retryable, it is rescheduled.

Multiple instances of this worker can be started.  Locks and queues are
stored in redis.  Use `lein test :redis` to run the tests.