Authenticate API through Surf Conext.
This change provides authentication and authorization of the HTTP API.

Clients calling the API endpoints should provide a valid Bearer token
from SURF Conext. The token is validated using the SURF Conext
introspection endpoint and if valid, the corresponding oin and
schac-home are retrieved from the clients config.

The clients config is read from a json-formatted config file (this
replaces the previous oin-mapping json file). If no configuration is
found for the token, or if no valid token is provided, the requests
are denied.

Since we now don't pass in insitution-schac-home directly, this also
updates the CLI tooling to take a client-id instead of a
schac-home. The client-id is used to fetch the corresponding
schac-home and institution-oin from the clients config.