Fix ClassNotFoundException
This is a workaround for a problem with an uknown root cause.

When excluding the clojurescript dependency in strgen, I cannot load
the repl or run the tests cause of the following error:

Execution error (ClassNotFoundException) at jdk.internal.loader.BuiltinClassLoader/loadClass (BuiltinClassLoader.java:641).
clojure.tools.reader.reader_types.Reader

This patch re-introduces the clojurescript dependency.

I added a ticket to fix the root problem later:
https://trello.com/c/cPZaqOeg/79