Narrow down detailed fieldsOfStudy for RIO
The OOAPI spec defines a fieldsOfStudy property which is an ISCED-F
string with 4 digits (a detailed code). The RIO API takes an ISCED
property of a number with range 0-999 (a narrow code). You can convert
from detailed to narrow by removing the last digit.

See also ISCED FIELDS OF EDUCATION AND TRAINING 2013 (ISCED-F 2013) -
Appendix I. ISCED fields of education and training.

http://uis.unesco.org/sites/default/files/documents/isced-fields-of-education-and-training-2013-en.pdf