Upgrade docker container to openjdk 17 (#318)
To support `UseContainerSupport` (set by default) and reliably use
`MaxRAMPercentage`.

When running in a (docker) container use the memory limit of the runtime
environment / cgroup and set environment variable `JAVA_TOOL_OPTIONS` to
`-XX:MaxRAMPercentage=75` to make sure the JVM uses at least 75% of the
allocated memory.  Note that 75% is a safe setting.